refactor(eval): fold online-eval into EvalClient
Core sub-clients are scoped to a CLI namespace, not an AWS resource type:
HarnessClient spans harness, versions, endpoints, invoke, and exec. Online
evaluation configs live under `agentcore eval` alongside evaluators, so they
belong on EvalClient / CoreEvalClient rather than a sibling core.onlineEval.

- Merged OnlineEvalClient's methods into EvalClient; deleted core/onlineEval.tsx.
- Moved CreateOnlineEvalInput / UpdateOnlineEvalInput into eval/types.tsx and
  the method signatures into CoreEvalClient; deleted online-eval/types.tsx.
- Dropped `onlineEval` from Core and CoreClient; handlers call core.eval.*.
- Folded TestOnlineEvalClient into TestEvalClient behind a shared record()
  helper.

Pure restructuring: no request shape changed, so every recorded fixture and
golden file still matches (513 tests green, unchanged).
