+++ src/handlers/types.tsx
@@ -1,7 +1,6 @@
 import type { CoreEvalClient } from "./eval/types.tsx";
 import type { CoreHarnessClient } from "./harness/types.tsx";
 import type { CoreIdentityClient } from "./identity/types.tsx";
-import type { CoreOnlineEvalClient } from "./eval/online-eval/types.tsx";
 import type { CoreRuntimeClient } from "./runtime/types.tsx";
 import type { Context } from "../router";
 import type { ProjectManager } from "./project/types.ts";
@@ -11,7 +10,6 @@ export interface Core {
   identity: CoreIdentityClient;
   runtime: CoreRuntimeClient;
   eval: CoreEvalClient;
-  onlineEval: CoreOnlineEvalClient;
   projectManager: ProjectManager;
 }
