+++ src/testing/index.tsx
@@ -7,7 +7,6 @@ export {
   TestHarnessClient,
   TestRuntimeClient,
   TestEvalClient,
-  TestOnlineEvalClient,
   type RecordedCall,
 } from "./TestCoreClient";
 export { StreamController } from "./StreamController";
